requirements.txt
gitpython>=3.1.0
pandas>=2.0.0
tree-sitter>=0.20.0
tree-sitter-python>=0.21.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
